Assembly: Put assembly groups in single file to avoid bloating.
