+++ src/Mod/Assembly/App/BomGroup.pyi
@@ -6,7 +6,7 @@ from Base.Metadata import export
 
 from App.DocumentObjectGroup import DocumentObjectGroup
 
-@export(Include="Mod/Assembly/App/BomGroup.h", Namespace="Assembly")
+@export(Include="Mod/Assembly/App/Groups.h", Namespace="Assembly")
 class BomGroup(DocumentObjectGroup):
     """
     This class is a group subclass for boms.
