+++ src/Mod/Assembly/App/Groups.h
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: LGPL-2.1-or-later
 /****************************************************************************
  *                                                                          *
- *   Copyright (c) 2023 Ondsel <development@ondsel.com>                     *
+ *   Copyright (c) 2024 Ondsel <development@ondsel.com>                     *
  *                                                                          *
  *   This file is part of FreeCAD.                                          *
  *                                                                          *
@@ -33,13 +33,69 @@
 namespace Assembly
 {
 
+class AssemblyExport BomGroup: public App::DocumentObjectGroup
+{
+    PROPERTY_HEADER_WITH_OVERRIDE(Assembly::BomGroup);
+
+public:
+    BomGroup() = default;
+    ~BomGroup() override = default;
+
+    PyObject* getPyObject() override;
+
+    /// returns the type name of the ViewProvider
+    const char* getViewProviderName() const override
+    {
+        return "AssemblyGui::ViewProviderBomGroup";
+    }
+};
+
+
+class AssemblyExport JointGroup: public App::DocumentObjectGroup
+{
+    PROPERTY_HEADER_WITH_OVERRIDE(Assembly::JointGroup);
+
+public:
+    JointGroup() = default;
+    ~JointGroup() override = default;
+
+    PyObject* getPyObject() override;
+
+    /// returns the type name of the ViewProvider
+    const char* getViewProviderName() const override
+    {
+        return "AssemblyGui::ViewProviderJointGroup";
+    }
+
+    std::vector<App::DocumentObject*> getJoints();
+};
+
+
+class AssemblyExport SimulationGroup: public App::DocumentObjectGroup
+{
+    PROPERTY_HEADER_WITH_OVERRIDE(Assembly::SimulationGroup);
+
+public:
+    SimulationGroup() = default;
+    ~SimulationGroup() override = default;
+
+    PyObject* getPyObject() override;
+
+    /// returns the type name of the ViewProvider
+    const char* getViewProviderName() const override
+    {
+        return "AssemblyGui::ViewProviderSimulationGroup";
+    }
+};
+
+
 class AssemblyExport ViewGroup: public App::DocumentObjectGroup
 {
     PROPERTY_HEADER_WITH_OVERRIDE(Assembly::ViewGroup);
 
 public:
-    ViewGroup();
-    ~ViewGroup() override;
+    ViewGroup() = default;
+    ~ViewGroup() override = default;
 
     PyObject* getPyObject() override;
 
@@ -49,6 +105,5 @@ public:
         return "AssemblyGui::ViewProviderViewGroup";
     }
 };
-
 
 }  // namespace Assembly
