+++ src/Mod/Assembly/App/JointGroup.pyi
@@ -6,7 +6,7 @@ from Base.Metadata import export
 
 from App.DocumentObjectGroup import DocumentObjectGroup
 
-@export(Include="Mod/Assembly/App/JointGroup.h", Namespace="Assembly")
+@export(Include="Mod/Assembly/App/Groups.h", Namespace="Assembly")
 class JointGroup(DocumentObjectGroup):
     """
     This class is a group subclass for joints.
