--- src/Mod/Assembly/App/SimulationGroup.cpp
@@ -1,52 +1,0 @@
-// SPDX-License-Identifier: LGPL-2.1-or-later
-/****************************************************************************
- *                                                                          *
- *   Copyright (c) 2024 Ondsel <development@ondsel.com>                     *
- *                                                                          *
- *   This file is part of FreeCAD.                                          *
- *                                                                          *
- *   FreeCAD is free software: you can redistribute it and/or modify it     *
- *   under the terms of the GNU Lesser General Public License as            *
- *   published by the Free Software Foundation, either version 2.1 of the   *
- *   License, or (at your option) any later version.                        *
- *                                                                          *
- *   FreeCAD is distributed in the hope that it will be useful, but         *
- *   WITHOUT ANY WARRANTY; without even the implied warranty of             *
- *   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU       *
- *   Lesser General Public License for more details.                        *
- *                                                                          *
- *   You should have received a copy of the GNU Lesser General Public       *
- *   License along with FreeCAD. If not, see                                *
- *   <https://www.gnu.org/licenses/>.                                       *
- *                                                                          *
- ***************************************************************************/
-
-
-#include <App/Application.h>
-#include <App/Document.h>
-#include <App/FeaturePythonPyImp.h>
-#include <App/PropertyPythonObject.h>
-#include <Base/Console.h>
-#include <Base/Tools.h>
-
-#include "SimulationGroup.h"
-#include "SimulationGroupPy.h"
-
-using namespace Assembly;
-
-
-PROPERTY_SOURCE(Assembly::SimulationGroup, App::DocumentObjectGroup)
-
-SimulationGroup::SimulationGroup()
-{}
-
-SimulationGroup::~SimulationGroup() = default;
-
-PyObject* SimulationGroup::getPyObject()
-{
-    if (PythonObject.is(Py::_None())) {
-        // ref counter is set to 1
-        PythonObject = Py::Object(new SimulationGroupPy(this), true);
-    }
-    return Py::new_reference_to(PythonObject);
-}
