--- src/Mod/Assembly/App/SimulationGroup.h
@@ -1,54 +1,0 @@
-// SPDX-License-Identifier: LGPL-2.1-or-later
-/****************************************************************************
- *                                                                          *
- *   Copyright (c) 2024 Ondsel <development@ondsel.com>                     *
- *                                                                          *
- *   This file is part of FreeCAD.                                          *
- *                                                                          *
- *   FreeCAD is free software: you can redistribute it and/or modify it     *
- *   under the terms of the GNU Lesser General Public License as            *
- *   published by the Free Software Foundation, either version 2.1 of the   *
- *   License, or (at your option) any later version.                        *
- *                                                                          *
- *   FreeCAD is distributed in the hope that it will be useful, but         *
- *   WITHOUT ANY WARRANTY; without even the implied warranty of             *
- *   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU       *
- *   Lesser General Public License for more details.                        *
- *                                                                          *
- *   You should have received a copy of the GNU Lesser General Public       *
- *   License along with FreeCAD. If not, see                                *
- *   <https://www.gnu.org/licenses/>.                                       *
- *                                                                          *
- ***************************************************************************/
-
-
-#pragma once
-
-#include <Mod/Assembly/AssemblyGlobal.h>
-
-#include <App/DocumentObjectGroup.h>
-#include <App/PropertyLinks.h>
-
-
-namespace Assembly
-{
-
-class AssemblyExport SimulationGroup: public App::DocumentObjectGroup
-{
-    PROPERTY_HEADER_WITH_OVERRIDE(Assembly::SimulationGroup);
-
-public:
-    SimulationGroup();
-    ~SimulationGroup() override;
-
-    PyObject* getPyObject() override;
-
-    /// returns the type name of the ViewProvider
-    const char* getViewProviderName() const override
-    {
-        return "AssemblyGui::ViewProviderSimulationGroup";
-    }
-};
-
-
-}  // namespace Assembly
