+++ src/Mod/Assembly/App/SimulationGroup.pyi
@@ -6,7 +6,7 @@ from Base.Metadata import export
 
 from App.DocumentObjectGroup import DocumentObjectGroup
 
-@export(Include="Mod/Assembly/App/SimulationGroup.h", Namespace="Assembly")
+@export(Include="Mod/Assembly/App/Groups.h", Namespace="Assembly")
 class SimulationGroup(DocumentObjectGroup):
     """
     This class is a group subclass for joints.
