+++ src/Mod/Assembly/Gui/AppAssemblyGui.cpp
@@ -30,10 +30,7 @@
 #include "ViewProviderAssembly.h"
 #include "ViewProviderAssemblyLink.h"
 #include "ViewProviderBom.h"
-#include "ViewProviderBomGroup.h"
-#include "ViewProviderJointGroup.h"
-#include "ViewProviderViewGroup.h"
-#include "ViewProviderSimulationGroup.h"
+#include "ViewProviderGroups.h"
 
 namespace AssemblyGui
 {
@@ -64,6 +61,7 @@ PyMOD_INIT_FUNC(AssemblyGui)
     AssemblyGui::ViewProviderAssembly::init();
     AssemblyGui::ViewProviderAssemblyLink::init();
     AssemblyGui::ViewProviderBom::init();
+    AssemblyGui::ViewProviderGroupBase::init();
     AssemblyGui::ViewProviderBomGroup::init();
     AssemblyGui::ViewProviderJointGroup::init();
     AssemblyGui::ViewProviderViewGroup::init();
