+++ src/Mod/Assembly/Gui/ViewProviderAssembly.cpp
@@ -71,9 +71,7 @@
 #include <Mod/Assembly/App/AssemblyLink.h>
 #include <Mod/Assembly/App/AssemblyObject.h>
 #include <Mod/Assembly/App/AssemblyUtils.h>
-#include <Mod/Assembly/App/JointGroup.h>
-#include <Mod/Assembly/App/ViewGroup.h>
-#include <Mod/Assembly/App/BomGroup.h>
+#include <Mod/Assembly/App/Groups.h>
 #include <Mod/PartDesign/App/Body.h>
 
 #include "TaskAssemblyMessages.h"
